Refactor JUnit/TestNGTestClassDetector
